Added vcpkg support for Windows building
- changed to use zlib/bzip2/lzma library names from vcpkg
- renamed XP build configuration so vcpkg recognises it as Release
- upgraded platform toolset to current 10.0.16299.0
- changed XP release to use 8.x WDK for WinUSB
- fixed Windows x64 build warnings, and restored bzip2 support

+++ src/BitBuffer.cpp
@@ -90,7 +90,7 @@ int BitBuffer::tell_index() const
 bool BitBuffer::seek_index (int index)
 {
 	if (index < 0)
-		index += m_indexes.size();
+		index += static_cast<int>(m_indexes.size());
 
 	if (index < 0 || index >= static_cast<int>(m_indexes.size()))
 		return false;
